+++ packages/xstate-inspect/src/browser.ts
@@ -72,7 +72,8 @@ const defaultInspectorOptions = {
     globalThis.__xstate__ = devTools;
     return devTools;
   },
-  serialize: undefined
+  serialize: undefined,
+  targetWindow: undefined
 };
 
 const getFinalOptions = (options?: Partial<InspectorOptions>) => {
@@ -88,9 +89,15 @@ const getFinalOptions = (options?: Partial<InspectorOptions>) => {
 const patchedInterpreters = new Set<AnyInterpreter>();
 
 export function inspect(options?: InspectorOptions): Inspector | undefined {
-  const { iframe, url, devTools } = getFinalOptions(options);
+  const finalOptions = getFinalOptions(options);
+  const { iframe, url, devTools } = finalOptions;
 
-  if (iframe === null) {
+  if (options?.targetWindow === null) {
+    throw new Error('Received a nullable `targetWindow`.');
+  }
+  let targetWindow: Window | null | undefined = finalOptions.targetWindow;
+
+  if (iframe === null && !targetWindow) {
     console.warn(
       'No suitable <iframe> found to embed the inspector. Please pass an <iframe> element to `inspect(iframe)` or create an <iframe data-xstate></iframe> element.'
     );
@@ -106,7 +113,6 @@ export function inspect(options?: InspectorOptions): Inspector | undefined {
     listeners.forEach((listener) => listener.next?.(state));
   });
 
-  let targetWindow: Window | null | undefined;
   let client: Pick<ActorRef<any>, 'send'>;
 
   const messageHandler = (event: MessageEvent<unknown>) => {
@@ -206,7 +212,7 @@ export function inspect(options?: InspectorOptions): Inspector | undefined {
     });
 
     iframe.setAttribute('src', String(url));
-  } else {
+  } else if (!targetWindow) {
     targetWindow = window.open(String(url), 'xstateinspector');
   }