+++ scripts/jest-utils/setup.d.ts
@@ -1,0 +1,9 @@
+import 'jest';
+
+declare global {
+  namespace jest {
+    interface Matchers<R, T> {
+      toMatchMockCallsInlineSnapshot(snapshot?: string): R;
+    }
+  }
+}